feat: refactor

+++ src/day_1/solution.rs
@@ -1,9 +1,9 @@
 use super::super::utils::read_one_per_line::read_one_per_line;
 
-pub fn solution() -> (usize, usize) {
+pub fn solution() -> (u64, u64) {
     let part_one_ans: usize = calc("./src/day_1/part1.txt", 2);
     let part_two_ans: usize = calc("./src/day_1/part2.txt", 4);
-    (part_one_ans, part_two_ans)
+    (part_one_ans as u64, part_two_ans as u64)
 }
 
 fn calc(path: &str, window: usize) -> usize {
